+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 Versions and bullets are arranged chronologically from latest to oldest.
 
-## Unreleased
+## v1.2.0
 
+- Implements GoogleAnalyticsPageViewService.
 - Uses $wgRelatedArticlesCardLimit for RelatedArticles.
 - Shows links on a page as RelatedArticles.
-- Implements GoogleAnalyticsPageViewService.
 
 ## v1.1.1